Fix segfault due to empty directories
With empty directories, two issues caused a segfault:
The linked list that contains the list of the files was freed
anyway--even if empty.
When the name of the directory had to be written, this name was obtained
through the specfications of the first contained file, which does not
exist in this case.