Fix superfluous newlines written to stdout
On error, do not write a newline to standard ouput. A lone newline is
used to separate the content of two directories written to stdout.
Thus, on error, there is no reason for a newline to appear. Even less on
stdout. This results in no visual 'separation' between the error line
and the line above it.