Fix segfault due to unavailable uid or gid entry
The structure returned by getpwuid() or getgrgid() is not checked for a
null. If the entry is not found, this results in a segmentation fault
when trying to calculate the lentgth of the string contained in the
structure.

This is solved by storing diretly either the string of the username or
the string of the uid.