Fix unfollowed arguments symlinks
Consequently to commit d68bc9b, for symlinks passed as arguments, the
link is followed only if it is specified, i.e., the path is ended by
'/'. There is no need to not follow the link in outputs different than
long list.

Now, the link's content is always outputted if a symlink is passed as
argument except in long list, if not specified.