Fix segfault on arguments error message
An error message, outputted if a file passed as argument is not valid,
causes a segfault when trying to read the string pointed to by
'filename'. This is a consequence to 7939119, where the appropriate
changes have not been done.

This is solved by always initializing the pointer 'filename' and
checking it directly for null.