Fix hyphen not detected as file name
A file who's name is just an hyphen and is passed as argument before
the separator is just ingored. Or, if a file name is composed of more
than two hyphens and is passed right after the options, it is equivalent
to the the separator. Specifying an hyphen as first option makes the
option considered as a file.

Checking exactly for the separator solves this issues.